edge_buffer now uses 32-bit indexing.

+++ fwdpp/ts/recording/edge_buffer.hpp
@@ -29,7 +29,7 @@ namespace fwdpp
             }
         };
 
-        using edge_buffer = nested_forward_lists<birth_data, std::int64_t, -1>;
+        using edge_buffer = nested_forward_lists<birth_data, std::int32_t, -1>;
 
         // Below are functions for liftover of an edge buffer
         // to a table collection
